@@ -17,8 +17,8 @@ module GovukContentModels
       private
 
       def make_record_action_noises(new_action, action_name)
-        NoisyWorkflow.make_noise(new_action).deliver
-        NoisyWorkflow.request_fact_check(new_action).deliver if action_name.to_s == "send_fact_check"
+        NoisyWorkflow.make_noise(new_action).deliver_now
+        NoisyWorkflow.request_fact_check(new_action).deliver_now if action_name.to_s == "send_fact_check"
       end
     end
   end
