@@ -8,9 +8,9 @@ class ActionDispatch::IntegrationTest
   include Capybara::DSL
 
   teardown do
-    DatabaseCleaner.clean
     Capybara.reset_sessions!    # Forget the (simulated) browser state
     Capybara.use_default_driver # Revert Capybara.current_driver to Capybara.default_driver
+    GDS::SSO.test_user = nil
   end
 
   def setup_users
