@@ -83,7 +83,7 @@ class EditionTest < ActiveSupport::TestCase
     # simulate validation error causing failure to publish anonymously
     edition.parts.first.update_attribute(:body, "[register your vehicle](registering-an-imported-vehicle)")
 
-    exception = assert_raises(StateMachine::InvalidTransition) { edition.publish_anonymously! }
+    exception = assert_raises(StateMachines::InvalidTransition) { edition.publish_anonymously! }
     assert_match "Cannot transition state via :publish from :ready (Reason(s): Parts", exception.message
     assert_match "Internal links must start with a forward slash", exception.message
   end
